requirements.txt
# Video download
yt-dlp>=2024.12.0

# Transcript extraction
youtube-transcript-api>=0.6.3

# Whisper fallback
openai-whisper>=20240930

# Translation + TTS (Gemini)
anthropic>=0.40.0
google-genai>=1.0.0

# TTS fallback
edge-tts>=6.1.0

# Audio processing
pydub>=0.25.1

# Web UI
gradio>=5.0.0

# Configuration
python-dotenv>=1.0.0
